requester: attach full error data to exceptions
Before these changes, users had to use the `suppress_http_errors` configuration option to access the full response data whenever the Plaid API returned an error status code. This made it impossible for consumers to both use exception handling to catch API call failures and then use the full response to take some corrective action.

This pull makes it so that all exceptions resulting from HTTP errors contain the fully decoded response JSON, allowing consumers to have the best of both worlds. Further, it makes it so that the response status code is an additional attribute available on exception instances.