client: Add webhook support to connect_update
This adds an explicit options parameter to connect_update, which allows consumers to update the connect account webhook URLs in a reasonable, documented way.

Previously, you had to do this by “tricking” the client into sending options data along by encoding it yourself and sticking the information into the login dictionary like so:

```python
    response = plaid.connect_update(login={
        'options': json.dumps({'webhook': webhook_url})
    }).json()
```

These changes make it so that the login information is explicitly not required and alleviate consumers from having to encode the options dict like many of the other plaid.client.Client methods.