Fix broken test suite (#68)
*  gitignore: add .cache directory

* travis: remove pre-release python 3.5 versions

* travis: use the project make script

* requirements: bump pytest reqs for py3.5 compat

The 3.x branch of pytest drops support for <3.3 which, according, to the travis file, this project still supports. As a consequence of that, we can only go up to pytest 2.9.2.

* setup.py: add python 3.5 trove classifier

This project is tested in that environment.

* client tests: correct ResourceNotFound testing

test_ResourceNotFound_connect was checking for a ResourceNotFoundError on an operation that should actually succeed. This modifies that operation so that it will fail as expected.